Phase 3 (suite): C4244 -- shims restants + double/float explicites
- 3 nouveaux fichiers avec le meme shim de portabilite deja identifie
  et retire de 7 autres fichiers (voir commit precedent) :
  ModifierObject3DSinusDisplace.cpp (sinf), Media3DView.cpp (atan2f),
  EquationList.cpp (sinf/cosf/sqrtf). Un espacement different (tabulation
  au lieu d'espace) les avait fait passer entre les mailles du grep
  precedent.

- Texture3D.cpp, Media3DView.cpp : cast (float) explicite sur des
  calculs trigonometriques dont l'argument est reellement en double
  precision (_frameDate est un double dans toute la hierarchie
  ProcessPreview) -- contrairement aux shims, ici le calcul en double
  est correct, seul le resultat final assigne a un membre float a
  besoin d'un cast explicite. Media3DView.cpp utilise cosf/sinf a la
  place quand l'argument est deja float (ffi).

- Object3DLWOResource.cpp, Object3DMultiplier.cpp : cast (float)
  explicite lors du passage d'un double (_frameDate) a des fonctions
  attendant un float (UpdateToFrame, SetFrameDate).

Warnings AzurVeda restants sous /W4 : 80 -> 66 (tous C4244).

Build revalide (exit 0, aucune erreur).

+++ VedaLib3DEngine/Object3DLWOResource.cpp
@@ -205,7 +205,7 @@ void	Object3DLWOResource::RenderObject( double _frameDate,
 				(((PackObjectReference *)(pCell->GetManagedObject()))->GetObjectPointer());
 			if(pTexture)
 			{
-				pTexture->UpdateToFrame(_frameDate);
+				pTexture->UpdateToFrame((float)_frameDate);
 				pVMTexture=	pTexture->GetVirtualMachineTexture();
 			}
 			pCell = pCell->GetNextBrotherCell();
